Data_pipeline/requirements.txt
apache-airflow==2.6.0
pandas
scikit-learn
fairlearn
matplotlib
python-dotenv
google-cloud-storage
requests
apache-airflow-providers-google
pytest
flake8
black
google-cloud-logging
google-cloud-bigquery
google-auth>=2.20.0
requests>=2.28.0
urllib3>=1.26.15
nltk
textblob
torch
transformers
datasets
# apache-airflow[gcp]==2.6.0
# apache-beam[gcp]==2.46.0
# pytest
# pandas
# google-cloud-storage==2.9.0
# google-cloud-bigquery==3.10.0
# google-cloud-logging==3.5.0
# dvc[gdrive]
# fairlearn
# scikit-learn
# google-cloud-dlp==3.21.0
